Project structure update
Merged "service.py" "randomData.py" files and moved to "utils" folder. Banner from config file moved to "utils" folder and named as "banner.py".
