+++ conf/config.py
@@ -1,29 +1,8 @@
-from colorama import Fore
-
 # set default value of default_country_code
 # only with numbers (without + sign)
 # Example:
 # default_country_code = 7
 default_country_code = None
-
-# FUCK MY BRAIN AND THIS SHIT
-banner=f"""
-{Fore.CYAN} (\__.-. |      {Fore.RED} ██████╗  ██████╗ ███╗   ███╗██████╗ ██╗██╗      █████╗{Fore.RESET}
-{Fore.CYAN}==== ===_]+     {Fore.RED} ██╔══██╗██╔═══██╗████╗ ████║██╔══██╗██║██║     ██╔══██╗{Fore.RESET}
-  {Fore.CYAN}```-   |      {Fore.RED} ██████╔╝██║   ██║██╔████╔██║██████╔╝██║██║     ███████║{Fore.RESET}
-      {Fore.CYAN}`.        {Fore.RED} ██╔══██╗██║   ██║██║╚██╔╝██║██╔══██╗██║██║     ██╔══██║{Fore.RESET}
-        {Fore.CYAN}` - >-> {Fore.RED} ██████╔╝╚██████╔╝██║ ╚═╝ ██║██████╔╝██║███████╗██║  ██║{Fore.RESET}
-                {Fore.RED} ╚═════╝{Fore.YELLOW}^ {Fore.RED}╚═════╝ ╚═╝{Fore.YELLOW} `\  {Fore.RED}╚═╝╚═════╝ ╚═╝╚══════╝╚═╝  ╚═╝
-{Fore.YELLOW}               __ /   (    <    /   )  \___
-- ------===;;;'====------------------===;;;===----- -  -
-                  \/  ~"~"~"~"~"~\~"~)~"/
-                  (_ (   \  (     >    \)
-                   \_( _ <         >_>'
-                      ~ `-i' ::>|--"
-                          I;|.|.|       
-                         <|i::|i|`.
-                        (` ^'"`-' ")
-                     ----`---------'---{Fore.RESET}"""
 
 # Random data for request
 emails = [
