+++ utils/banner.py
@@ -1,0 +1,21 @@
+from colorama import Fore
+
+
+# FUCK MY BRAIN AND THIS SHIT
+banner=f"""
+{Fore.CYAN} (\__.-. |      {Fore.RED} ██████╗  ██████╗ ███╗   ███╗██████╗ ██╗██╗      █████╗{Fore.RESET}
+{Fore.CYAN}==== ===_]+     {Fore.RED} ██╔══██╗██╔═══██╗████╗ ████║██╔══██╗██║██║     ██╔══██╗{Fore.RESET}
+  {Fore.CYAN}```-   |      {Fore.RED} ██████╔╝██║   ██║██╔████╔██║██████╔╝██║██║     ███████║{Fore.RESET}
+      {Fore.CYAN}`.        {Fore.RED} ██╔══██╗██║   ██║██║╚██╔╝██║██╔══██╗██║██║     ██╔══██║{Fore.RESET}
+        {Fore.CYAN}` - >-> {Fore.RED} ██████╔╝╚██████╔╝██║ ╚═╝ ██║██████╔╝██║███████╗██║  ██║{Fore.RESET}
+                {Fore.RED} ╚═════╝{Fore.YELLOW}^ {Fore.RED}╚═════╝ ╚═╝{Fore.YELLOW} `\  {Fore.RED}╚═╝╚═════╝ ╚═╝╚══════╝╚═╝  ╚═╝
+{Fore.YELLOW}               __ /   (    <    /   )  \___
+- ------===;;;'====------------------===;;;===----- -  -
+                  \/  ~"~"~"~"~"~\~"~)~"/
+                  (_ (   \  (     >    \)
+                   \_( _ <         >_>'
+                      ~ `-i' ::>|--"
+                          I;|.|.|
+                         <|i::|i|`.
+                        (` ^'"`-' ")
+                     ----`---------'---{Fore.RESET}"""
